refactor: move diff view, rollback, file ops into their own modules
