+++ src/render.rs
@@ -714,569 +714,10 @@ impl Kyde {
 
 
 
-    /// Side-by-side diff = two editors in one rounded island: left is the read-only base
-    /// (HEAD/index), right is the editable working copy (live-saved). A draggable divider
-    /// sets the 50/50 split. Both syntax-highlight when the language pack is installed.
-    /// Full-screen Show-Diff view in the MAIN window (rollback / push diff). A header with a
-    /// Back button + the file path over the inline `render_diff`, so it reuses every editor
-    /// feature — find (⌘F), divider drag, scrollbars, change navigation. Escape / Back exits.
-    fn render_diff_view(
-        &mut self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        let title: SharedString = self
-            .diff_path
-            .as_ref()
-            .map(|p| p.to_string_lossy().into_owned())
-            .unwrap_or_else(|| "Diff".to_string())
-            .into();
-        let back = div()
-            .id("diff-back")
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_1()
-            .px_2()
-            .py_0p5()
-            .rounded_md()
-            .cursor_pointer()
-            .text_color(t.secondary_text)
-            .hover(|s| s.bg(t.bg_mid).text_color(t.text))
-            .child("‹ Back")
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.close_diff_view(cx)),
-            );
-        let header = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_2()
-            .flex_none()
-            .px_1()
-            .pb(px(theme::FRAME_GAP))
-            .font_family(ui)
-            .text_size(px(t.ui_font_size))
-            .child(back)
-            .child(div().min_w_0().truncate().text_color(t.text).child(title));
-        let mut col = div()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .min_h_0()
-            .min_w_0()
-            .child(header);
-        // The find/replace bar (⌘F / ⌘R), targeting whichever diff pane is focused.
-        if self.find_open {
-            col = col.child(self.render_find_bar(ui, cx));
-        }
-        col.child(self.render_diff(ui, fs, Some(window), cx))
-            .into_any_element()
-    }
 
-    /// IntelliJ-style side-by-side diff: aligned rows, with a center gutter showing the old
-    /// and new line numbers, a `»` chevron (revert the hunk) and a checkbox (stage it).
-    pub(crate) fn render_diff(
-        &mut self,
-        _ui: &'static str,
-        fs: gpui::Pixels,
-        mut window: Option<&mut Window>,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        let island = || {
-            div()
-                .flex_1()
-                .min_w_0()
-                .h_full()
-                .bg(t.main_bg)
-                .rounded(px(theme::ISLAND_RADIUS))
-                .overflow_hidden()
-                .font_family(theme::font::FAMILY)
-                .text_size(fs)
-                .text_color(t.text)
-        };
 
-        // The Diff modal window renders the same `diff_left`/`diff_right` editors; rendering one
-        // editor entity in two windows desyncs scroll + garbles layout. So while the modal is
-        // open, the INLINE diff (the only caller that passes `Some(window)`) yields the panes to
-        // it and shows a placeholder. The modal itself passes `None`, so it renders normally.
-        if window.is_some() && self.diff_modal_open {
-            return island()
-                .flex()
-                .items_center()
-                .justify_center()
-                .text_color(t.line_number)
-                .child("Viewing diff in window…")
-                .into_any_element();
-        }
 
-        // Image file selected → preview it centered + scaled (same as Browse), not a text diff.
-        if let Some(rel) = self.diff_image.clone() {
-            let abs = self.repo_root.as_ref().map(|r| r.join(&rel)).unwrap_or(rel);
-            return island()
-                .id("diff-image-scroll")
-                .overflow_scroll()
-                .flex()
-                .items_center()
-                .justify_center()
-                .p_4()
-                .child(img(abs).max_w_full().max_h_full())
-                .into_any_element();
-        }
 
-        let Some(d) = self
-            .current_diff
-            .as_ref()
-            .filter(|_| self.diff_path.is_some())
-        else {
-            return island()
-                .flex()
-                .justify_center()
-                .items_center()
-                .text_color(t.line_number)
-                .child("Select a file")
-                .into_any_element();
-        };
-
-        // Center gutter: a `»` (revert this hunk) on each hunk's first row, sharing
-        // `diff_scroll` so it tracks the editors. Chevrons are positioned ABSOLUTELY at
-        // `row * row_h` inside a fixed-height column — a flex column of per-row divs let
-        // empty rows collapse (gpui ignores `.h()` on a childless div), which bunched every
-        // chevron toward the top instead of onto its hunk row.
-        let row_h = px(editor::line_height_px());
-        let rows = aligned_rows(d);
-        let total_h = row_h * rows.len() as f32;
-        // Read-only diffs (push view) show a committed change with no working-tree edit,
-        // so there's nothing to revert — drop the gutter chevrons.
-        let chevrons: Vec<gpui::AnyElement> = if self.diff_readonly {
-            Vec::new()
-        } else {
-            rows.iter()
-                .enumerate()
-                .filter_map(|(i, r)| {
-                    let hi = r.hunk_start.then_some(r.hunk).flatten()?;
-                    Some(
-                        // Position the row's line box exactly over the editor line (`line_height`
-                        // = `row_h`), so the `»` baselines with the line's text instead of being
-                        // half-centered a few px high.
-                        div()
-                            .absolute()
-                            .top(row_h * i as f32)
-                            .left_0()
-                            .right_0()
-                            .h(row_h)
-                            .flex()
-                            .flex_row()
-                            .items_center()
-                            .justify_center()
-                            .child(
-                                // Flex-center the glyph in a fixed box so it's truly centered
-                                // (a line-height hack left it sitting low).
-                                div()
-                                    .id(SharedString::from(format!("revert-{hi}")))
-                                    .flex()
-                                    .items_center()
-                                    .justify_center()
-                                    .size(px(24.0))
-                                    .rounded_sm()
-                                    .text_size(px(20.0))
-                                    .line_height(px(20.0))
-                                    .text_color(t.line_number)
-                                    .hover(|s| s.bg(t.bg_light).text_color(t.primary))
-                                    .cursor_pointer()
-                                    .child("»")
-                                    .on_mouse_down(
-                                        MouseButton::Left,
-                                        cx.listener(move |this, _e, _w, cx| {
-                                            this.diff_revert_hunk(hi, cx)
-                                        }),
-                                    ),
-                            )
-                            .into_any_element(),
-                    )
-                })
-                .collect()
-        };
-
-        // Pane = shared VERTICAL scroll (`diff_scroll`, keeps the two sides' rows aligned)
-        // wrapping an INDEPENDENT horizontal scroll around a content-width editor, so long
-        // lines scroll sideways per pane without breaking row alignment.
-        let frac = self.diff_split.clamp(0.15, 0.85);
-        let lw = self.diff_left.read(cx).content_width();
-        let rw = self.diff_right.read(cx).content_width();
-        let pane_scroll =
-            |id: &'static str, scroll: &ScrollHandle, w: f32, editor: gpui::AnyElement| {
-                div()
-                    .id(id)
-                    .h_full()
-                    .w_full()
-                    .overflow_scroll()
-                    .track_scroll(scroll)
-                    .child(div().w(px(w)).child(editor))
-            };
-        let left_inner = pane_scroll(
-            "diff-left-scroll",
-            &self.diff_scroll,
-            lw,
-            self.diff_left.clone().into_any_element(),
-        );
-        let right_inner = pane_scroll(
-            "diff-right-scroll",
-            &self.diff_scroll,
-            rw,
-            self.diff_right.clone().into_any_element(),
-        );
-        // One shared horizontal scrollbar driven by `diff_scroll`; its travel is the wider
-        // pane's content. Placed on the island (full width) below.
-        let h_bar = self.diff_hscrollbar(
-            &self.diff_scroll.clone(),
-            lw.max(rw),
-            SbView::DiffLeftH,
-            window.as_deref_mut(),
-            cx,
-        );
-        // New file (empty left) or deleted file (empty right): show ONLY the populated side,
-        // full-width. A side-by-side with one empty pane is noise — and the empty pane drives
-        // the shared scroll handle's bounds to ~0, which blanks the viewport-culled editor on a
-        // large file. Full-width, the surviving editor owns the layout and paints normally.
-        let left_empty = self.diff_left.read(cx).text().is_empty();
-        let right_empty = self.diff_right.read(cx).text().is_empty();
-        if left_empty != right_empty {
-            let inner = if left_empty { right_inner } else { left_inner };
-            let scrollbar = self.diff_vscrollbar(total_h, window.as_deref_mut(), cx);
-            return island()
-                .relative()
-                .flex()
-                .flex_row()
-                .child(div().relative().flex_1().min_w_0().h_full().child(inner))
-                .children(h_bar)
-                .children(scrollbar)
-                .children(self.render_diff_nav(cx))
-                .into_any_element();
-        }
-
-        // Left pane width: when we know the viewport (the inline diff, the only resizable
-        // caller — the modal passes `None`), size it to an explicit pixel width from the SAME
-        // formula the resize handler inverts, and let the right pane flex to fill the rest. The
-        // modal falls back to proportional flex_basis (its divider isn't dragged through the
-        // main-window handler). Both keep `min_w_0` so neither pane can be pushed off-window.
-        let left_px = window.as_deref().map(|w| {
-            (frac * (full_island_w(f32::from(w.viewport_size().width)) - DIFF_GUTTER_W)).max(40.0)
-        });
-        let left = div().relative().min_w_0().h_full();
-        let left = match left_px {
-            Some(w) => left.w(px(w)).flex_none(),
-            None => left.flex_basis(gpui::relative(frac)).flex_shrink(),
-        }
-        .child(left_inner);
-        let right = div().relative().min_w_0().h_full();
-        let right = match left_px {
-            Some(_) => right.flex_1(),
-            None => right.flex_basis(gpui::relative(1.0 - frac)).flex_shrink(),
-        }
-        .child(right_inner);
-        // The gutter (chevrons) shares the editors' vertical scroll by translating its content
-        // by the SAME offset; it also doubles as the draggable divider (drag to resize the
-        // split). Clicks on a `»` still revert their hunk (chevrons are children).
-        let scroll_y = self.diff_scroll.offset().y;
-        let gutter = div()
-            .id("diff-gutter")
-            .w(px(DIFF_GUTTER_W))
-            .flex_none()
-            .h_full()
-            .overflow_hidden()
-            .bg(t.diff_separator_bg)
-            // Thin divider lines on both edges so the center gutter reads as the
-            // pane separator (otherwise it blends into the panes at most font sizes).
-            .border_l(px(1.0))
-            .border_r(px(1.0))
-            .border_color(t.divider)
-            .cursor_col_resize()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, e: &gpui::MouseDownEvent, window, cx| {
-                    this.start_divider_drag(Divider::DiffPane, e.position, window);
-                    cx.notify();
-                }),
-            )
-            .child(
-                div()
-                    .relative()
-                    .w_full()
-                    .h(total_h)
-                    .top(scroll_y)
-                    .children(chevrons),
-            );
-
-        // Single vertical scrollbar overlaid on the right edge of the island; both panes
-        // share `diff_scroll`, so one bar drives the whole diff. `total_h` is the exact
-        // aligned-row content height (both panes are padded to it), so the bar is driven by
-        // that rather than the shared handle's `max_offset` — which reflects whichever pane
-        // painted last and is ~0 when one side is empty (e.g. an all-added/untracked file).
-        let scrollbar = self.diff_vscrollbar(total_h, window, cx);
-
-        island()
-            .relative()
-            .flex()
-            .flex_row()
-            .child(left)
-            .child(gutter)
-            .child(right)
-            .children(h_bar)
-            .children(scrollbar)
-            .children(self.render_diff_nav(cx))
-            .into_any_element()
-    }
-
-    /// A small floating control at the diff's top-right: the change count + prev/next arrows
-    /// that jump between hunks. Shown whenever the current diff has at least one change.
-    fn render_diff_nav(&self, cx: &mut Context<Self>) -> Option<gpui::AnyElement> {
-        let n = self
-            .current_diff
-            .as_ref()
-            .map(|d| d.hunks.len())
-            .unwrap_or(0);
-        // Nothing to navigate with a single (or zero) change — hide the control.
-        if n < 2 {
-            return None;
-        }
-        let t = theme::get();
-        let ui = theme::font::UI_FAMILY;
-        let label = format!("{n} changes");
-        let arrow = |id: &'static str, icon: &'static str, tip: &'static str, next: bool| {
-            div()
-                .id(id)
-                .size(px(20.0))
-                .flex()
-                .items_center()
-                .justify_center()
-                .rounded_md()
-                .cursor_pointer()
-                .hover(|s| s.bg(t.bg_light))
-                .tooltip(move |_w, cx| cx.new(|_| Tip(tip.into())).into())
-                .child(svg().path(icon).size(px(14.0)).text_color(t.secondary_text))
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(move |this, _e, _w, cx| {
-                        cx.stop_propagation();
-                        this.diff_nav_hunk(next, cx);
-                    }),
-                )
-        };
-        Some(
-            div()
-                .absolute()
-                .top(px(8.0))
-                // Clear the 12px vertical scrollbar gutter on the right edge.
-                .right(px(16.0))
-                .flex()
-                .flex_row()
-                .items_center()
-                .gap_1()
-                // Swallow clicks on the control (label + padding included) so they don't reach
-                // the editor underneath and start a text selection.
-                .on_mouse_down(MouseButton::Left, |_, _, cx| cx.stop_propagation())
-                .rounded_md()
-                .border_1()
-                .border_color(t.divider)
-                .bg(t.bg_mid)
-                .pl_2()
-                .pr_1()
-                .py_0p5()
-                .font_family(ui)
-                .text_size(px(t.ui_font_size))
-                .text_color(t.secondary_text)
-                .child(div().mr_1().child(label))
-                .child(arrow(
-                    "diff-prev",
-                    "icons/arrow-up.svg",
-                    "Previous change",
-                    false,
-                ))
-                .child(arrow(
-                    "diff-next",
-                    "icons/arrow-down.svg",
-                    "Next change",
-                    true,
-                ))
-                .into_any_element(),
-        )
-    }
-
-    /// A vertical scrollbar thumb overlaid on the right edge of the diff island, driven by
-    /// `diff_scroll` (both panes share it). Returns `None` when the diff fits. Absolutely
-    /// positioned — unlike `with_scrollbars` it doesn't need a concrete pane width, so it
-    /// works over the diff's flex 50/50 split. `window` (when `Some`, i.e. the inline commit
-    /// view) requests one settle frame so the bar appears on first paint, before any scroll;
-    /// the modal diff passes `None` and the bar shows after the first scroll/repaint.
-    fn diff_vscrollbar(
-        &mut self,
-        total_h: gpui::Pixels,
-        window: Option<&mut Window>,
-        cx: &mut Context<Self>,
-    ) -> Option<gpui::AnyElement> {
-        let t = theme::get();
-        let scroll = self.diff_scroll.clone();
-        // Viewport height is reliable from the shared handle (both panes are `h_full` in the
-        // same island, so they paint the same visible height). The scroll *distance* is the
-        // content height beyond the viewport — computed from `total_h`, not `max_offset`.
-        let vp_h = scroll.bounds().size.height;
-        let off = scroll.offset();
-        let max_scroll = (total_h - vp_h).max(px(0.0));
-        const BAR: f32 = 12.0;
-
-        // Scroll metrics are zero until the panes have painted, so the first frame after a
-        // file loads can't know whether a bar is needed. Track the painted dims and ask for
-        // one more frame when they change, so the bar settles in without a scroll/resize.
-        if let Some(window) = window {
-            let dims: ScrollDims = (px(0.0), px(0.0), vp_h, px(0.0), total_h);
-            if self.scroll_dims.get(&SbView::Diff) != Some(&dims) {
-                self.scroll_dims.insert(SbView::Diff, dims);
-                let entity = cx.entity();
-                window.on_next_frame(move |_, cx| entity.update(cx, |_, cx| cx.notify()));
-            }
-        }
-
-        if max_scroll <= px(1.0) {
-            return None; // fits — nothing to scroll
-        }
-
-        const END: f32 = 8.0;
-        const THUMB: f32 = 6.0;
-        let (thumb_h, top) = scrollbar_thumb(
-            f32::from(vp_h),
-            f32::from(max_scroll),
-            f32::from(off.y),
-            END,
-        );
-        let m = (BAR - THUMB) / 2.0;
-        let sc = scroll.clone();
-        let thumb = div()
-            .id("diff-sb-v")
-            .absolute()
-            .top(px(top))
-            .left(px(m))
-            .w(px(THUMB))
-            .h(px(thumb_h))
-            .rounded_full()
-            .bg(t.line_number)
-            .opacity(0.5)
-            .hover(|s| s.opacity(0.85))
-            .cursor_pointer()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(move |this, e: &gpui::MouseDownEvent, _w, cx| {
-                    // Don't let the editor/tree underneath also handle this — it would start a
-                    // text selection / move the caret behind the scrollbar.
-                    cx.stop_propagation();
-                    this.sb_drag = Some(crate::SbDrag {
-                        handle: sc.clone(),
-                        horizontal: false,
-                        start_cursor: f32::from(e.position.y),
-                        start_off: f32::from(sc.offset().y),
-                    });
-                    cx.notify();
-                }),
-            );
-        Some(
-            div()
-                .absolute()
-                .top_0()
-                .bottom_0()
-                .right_0()
-                .w(px(BAR))
-                // Swallow clicks anywhere in the scrollbar gutter (track included) so they never
-                // reach the editor underneath and start a text selection.
-                .on_mouse_down(MouseButton::Left, |_, _, cx| cx.stop_propagation())
-                .child(thumb)
-                .into_any_element(),
-        )
-    }
-
-    /// Horizontal scrollbar overlaid at the bottom of one diff pane, driven by `scroll` (that
-    /// pane's independent horizontal handle) and `content_w` (its widest line). Returns `None`
-    /// when the lines fit. The caller overlays it on a `relative` pane wrapper.
-    fn diff_hscrollbar(
-        &mut self,
-        scroll: &ScrollHandle,
-        content_w: f32,
-        view: SbView,
-        window: Option<&mut Window>,
-        cx: &mut Context<Self>,
-    ) -> Option<gpui::AnyElement> {
-        let t = theme::get();
-        let vp_w = scroll.bounds().size.width;
-        let off = scroll.offset();
-        let max_scroll = px(content_w) - vp_w;
-        const BAR: f32 = 12.0;
-        if let Some(window) = window {
-            let dims: ScrollDims = (vp_w, px(content_w), px(0.0), px(0.0), px(0.0));
-            if self.scroll_dims.get(&view) != Some(&dims) {
-                self.scroll_dims.insert(view, dims);
-                let entity = cx.entity();
-                window.on_next_frame(move |_, cx| entity.update(cx, |_, cx| cx.notify()));
-            }
-        }
-        if max_scroll <= px(1.0) {
-            return None; // lines fit — no horizontal scroll
-        }
-        const END: f32 = 8.0;
-        const THUMB: f32 = 6.0;
-        let (thumb_w, left) = scrollbar_thumb(
-            f32::from(vp_w),
-            f32::from(max_scroll),
-            f32::from(off.x),
-            END,
-        );
-        let m = (BAR - THUMB) / 2.0;
-        let sc = scroll.clone();
-        let id = if matches!(view, SbView::DiffLeftH) {
-            "diff-sb-h-l"
-        } else {
-            "diff-sb-h-r"
-        };
-        let thumb = div()
-            .id(id)
-            .absolute()
-            .left(px(left))
-            .top(px(m))
-            .h(px(THUMB))
-            .w(px(thumb_w))
-            .rounded_full()
-            .bg(t.line_number)
-            .opacity(0.5)
-            .hover(|s| s.opacity(0.85))
-            .cursor_pointer()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(move |this, e: &gpui::MouseDownEvent, _w, cx| {
-                    cx.stop_propagation();
-                    this.sb_drag = Some(crate::SbDrag {
-                        handle: sc.clone(),
-                        horizontal: true,
-                        start_cursor: f32::from(e.position.x),
-                        start_off: f32::from(sc.offset().x),
-                    });
-                    cx.notify();
-                }),
-            );
-        Some(
-            div()
-                .absolute()
-                .left_0()
-                .right_0()
-                .bottom_0()
-                .h(px(BAR))
-                .on_mouse_down(MouseButton::Left, |_, _, cx| cx.stop_propagation())
-                .child(thumb)
-                .into_any_element(),
-        )
-    }
 
 
     /// Shared tree row used by Browse, Commit, and Rollback — so items look identical
@@ -3438,24 +2879,6 @@ impl Kyde {
             .into_any_element()
     }
 
-    /// Floating "Show Diff" viewer over the Commit view (IntelliJ-style).
-    /// Show-Diff window body (own native window; titlebar shows the file path). Just the
-    /// side-by-side diff filling the window.
-    pub(crate) fn render_diff_body(&mut self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let ui = theme::font::UI_FAMILY;
-        let fs = px(theme::get().editor_font_size);
-        let t = theme::get();
-        div()
-            .size_full()
-            .flex()
-            .flex_col()
-            .bg(t.main_bg)
-            .font_family(ui)
-            .text_size(fs)
-            .text_color(t.text)
-            .child(self.render_diff(ui, fs, None, cx))
-            .into_any_element()
-    }
 
 
     /// "Create New Branch" dialog body (own native window). Name field (`branch_query`) +
@@ -3555,76 +2978,6 @@ impl Kyde {
             .into_any_element()
     }
 
-    fn render_delete_modal(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let Some((path, is_dir)) = self.delete_target.clone() else {
-            return div().into_any_element();
-        };
-        let name = path
-            .file_name()
-            .map(|n| n.to_string_lossy().into_owned())
-            .unwrap_or_else(|| path.to_string_lossy().into_owned());
-        let kind = if is_dir { "folder" } else { "file" };
-
-        let cancel = btn_secondary("delete-cancel", "Cancel")
-            .px_3()
-            .py_1p5()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| {
-                    this.delete_target = None;
-                    cx.notify();
-                }),
-            );
-        let confirm = div()
-            .id("delete-confirm")
-            .px_3()
-            .py_1p5()
-            .rounded_md()
-            .bg(t.status_deleted)
-            .text_color(t.primary_text)
-            .cursor_pointer()
-            .hover(|s| s.opacity(0.9))
-            .child("Delete")
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.do_delete(cx)),
-            );
-
-        let panel = div()
-            .w(px(420.0))
-            .flex()
-            .flex_col()
-            .gap_3()
-            .p_4()
-            .bg(t.frame_bg)
-            .border_1()
-            .border_color(t.divider)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .shadow_lg()
-            .font_family(ui)
-            .text_size(px(theme::get().ui_font_size + 1.0))
-            .occlude()
-            .child(div().text_color(t.text).child(format!("Delete {kind}?")))
-            .child(
-                div()
-                    .text_color(t.secondary_text)
-                    .text_size(px(theme::get().ui_font_size))
-                    .child(format!(
-                        "“{name}” will be permanently deleted from disk. This can't be undone."
-                    )),
-            )
-            .child(
-                div()
-                    .flex()
-                    .flex_row()
-                    .justify_end()
-                    .gap_2()
-                    .child(cancel)
-                    .child(confirm),
-            );
-        overlay(cx, true).child(panel).into_any_element()
-    }
 
     /// Body of the "Clear Data & Restart" confirmation modal window. Destructive: wipes the
     /// whole config dir and restarts.
@@ -3670,242 +3023,7 @@ impl Kyde {
             .into_any_element()
     }
 
-    /// New-file / rename modal: a single-line name input + Create/Rename & Cancel.
-    /// Enter confirms / Esc cancels via the "FileFinder" key context (the input is
-    /// single-line, so those keys bubble up to this wrapper).
-    fn render_name_prompt(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let Some(prompt) = self.name_prompt.clone() else {
-            return div().into_any_element();
-        };
-        let (title, action) = match &prompt {
-            NamePrompt::NewFile(_) => ("New file", "Create"),
-            NamePrompt::Rename(_) => ("Rename", "Rename"),
-        };
 
-        let cancel = btn_secondary("name-cancel", "Cancel")
-            .px_3()
-            .py_1p5()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, window, cx| this.cancel_name_prompt(window, cx)),
-            );
-        let confirm = btn_primary("name-confirm", action)
-            .px_3()
-            .py_1p5()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, window, cx| this.confirm_name_prompt(window, cx)),
-            );
-
-        let panel =
-            div()
-                .key_context("FileFinder")
-                .on_action(cx.listener(|this, _: &FinderConfirm, window, cx| {
-                    this.confirm_name_prompt(window, cx)
-                }))
-                .on_action(cx.listener(|this, _: &FinderClose, window, cx| {
-                    this.cancel_name_prompt(window, cx)
-                }))
-                .w(px(420.0))
-                .flex()
-                .flex_col()
-                .gap_3()
-                .p_4()
-                .bg(t.frame_bg)
-                .border_1()
-                .border_color(t.divider)
-                .rounded(px(theme::ISLAND_RADIUS))
-                .shadow_lg()
-                .font_family(ui)
-                .text_size(px(theme::get().ui_font_size + 1.0))
-                .occlude()
-                .child(div().text_color(t.text).child(title))
-                .child(
-                    div()
-                        .px_2()
-                        .py_1()
-                        .rounded_md()
-                        .bg(t.main_bg)
-                        .border_1()
-                        .border_color(t.divider)
-                        .child(self.name_input.clone()),
-                )
-                .child(
-                    div()
-                        .flex()
-                        .flex_row()
-                        .justify_end()
-                        .gap_2()
-                        .child(cancel)
-                        .child(confirm),
-                );
-        overlay(cx, true).child(panel).into_any_element()
-    }
-
-    /// Rollback Changes window body: checkbox tree of changes + Close/Rollback. Right-click a
-    /// row → View Diff. Hosted in its own native window (`ModalWindow`).
-    pub(crate) fn render_rollback_body(&mut self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let ui = theme::font::UI_FAMILY;
-        let fs = px(theme::get().ui_font_size);
-        let t = theme::get();
-        let root_name = self
-            .repo_root
-            .as_ref()
-            .and_then(|p| p.file_name())
-            .map(|s| s.to_string_lossy().into_owned())
-            .unwrap_or_else(|| "/".to_string());
-        // Same folder tree as the Commit view, with rollback checkboxes.
-        let mut visible = vec![tree::Row {
-            path: PathBuf::new(),
-            is_dir: true,
-            depth: 0,
-        }];
-        if self.commit_expanded.contains(&PathBuf::new()) {
-            for mut r in self.commit_tree.visible(&self.commit_expanded) {
-                r.depth += 1;
-                visible.push(r);
-            }
-        }
-        let rows: Vec<gpui::AnyElement> = visible
-            .into_iter()
-            .map(|r| {
-                let is_root = r.path.as_os_str().is_empty();
-                let checked = if r.is_dir {
-                    self.rollback_folder_all_checked(&r.path)
-                } else {
-                    self.rollback_checked.contains(&r.path)
-                };
-                let name_color = self
-                    .files
-                    .iter()
-                    .find(|f| f.path == r.path)
-                    .map(|f| status_color(f.status))
-                    .unwrap_or(t.text);
-                let name: SharedString = if is_root {
-                    root_name.clone().into()
-                } else {
-                    r.path
-                        .file_name()
-                        .map(|n| n.to_string_lossy().into_owned())
-                        .unwrap_or_default()
-                        .into()
-                };
-                let expanded = self.commit_expanded.contains(&r.path);
-                let is_dir = r.is_dir;
-                let (p_act, p_check, p_ctx) = (r.path.clone(), r.path.clone(), r.path.clone());
-                self.tree_row(
-                    cx,
-                    &r.path,
-                    is_dir,
-                    expanded,
-                    r.depth,
-                    false,
-                    name,
-                    name_color,
-                    Some(checked),
-                    // Click toggles the row's checkbox (folders expand); the diff is reached
-                    // by right-click → View Diff, not a plain click.
-                    move |this, _e, _w, cx| {
-                        if is_dir {
-                            this.toggle_commit_dir(p_act.clone(), cx);
-                        } else {
-                            this.toggle_rollback_check(p_act.clone(), false, cx);
-                        }
-                    },
-                    move |this, cx| this.toggle_rollback_check(p_check.clone(), is_dir, cx),
-                    move |this, pos, cx| {
-                        if !is_dir {
-                            if let Some(i) = this.files.iter().position(|f| f.path == p_ctx) {
-                                this.open_menu(pos, MenuTarget::RollbackFile(i), cx);
-                            }
-                        }
-                    },
-                )
-            })
-            .collect();
-
-        let count = self.files.len();
-
-        let list = div()
-            .id("rollback-list")
-            .overflow_y_scroll()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .p_2()
-            .child(
-                div()
-                    .px_1()
-                    .py_1()
-                    .text_color(t.line_number)
-                    .child(SharedString::from(format!("Changes  {count} files"))),
-            )
-            .children(rows);
-
-        let delete_added = self.rollback_delete_added;
-        let delete_row = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_2()
-            .px_3()
-            .py_2()
-            .text_color(t.text)
-            .cursor_pointer()
-            .child(checkbox_box(delete_added))
-            .child("Delete local copies of added files")
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| {
-                    this.rollback_delete_added = !this.rollback_delete_added;
-                    cx.notify();
-                }),
-            );
-
-        let footer = div()
-            .flex()
-            .flex_row()
-            .justify_end()
-            .gap_2()
-            .px_3()
-            .py_2()
-            .border_t_1()
-            .border_color(t.divider)
-            .child(
-                btn_secondary("rollback-close", "Close")
-                    .py_1p5()
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, _e, _w, cx| this.close_rollback_window(cx)),
-                    ),
-            )
-            .child(btn_primary("rollback", "Rollback").py_1p5().on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.do_rollback(cx)),
-            ));
-
-        // Fills its native window (the OS titlebar shows "Rollback Changes" + close button).
-        let mut panel = div()
-            .size_full()
-            .flex()
-            .flex_col()
-            .bg(t.panel_bg)
-            .font_family(ui)
-            .text_size(fs)
-            .text_color(t.text)
-            .child(list)
-            .child(delete_row)
-            .child(footer);
-        // Right-click → View Diff menu renders inside THIS window (its target is a RollbackFile).
-        if matches!(
-            self.context_menu.as_ref().map(|m| &m.target),
-            Some(MenuTarget::RollbackFile(_))
-        ) {
-            panel = panel.child(self.render_context_menu(cx));
-        }
-        panel.into_any_element()
-    }
 
     /// Font specimen viewer: each bundled family at every available weight, previewing a
     /// full sentence at a large size so weight differences are obvious.
